Basic Python server for testing streaming, plus streaming scripts
This is an initial commit of what I have working so far for streaming to
browsers from FFMPEG with the help of hls.js. The goal is to remove the
bash scripts which were used for testing and have operation of VHS dubbing
fully controlled from a web page using Python.

Note that the icon used is from artist "Lionti_design" on TeePublic and
is used without permission. This is a private system and I just needed
an icon for now. I will remove this when the repository is published
publicly:

  https://www.teepublic.com/de/aufkleber/1645308-vhs-tapes-icon?store_id=117652